Time: 737 ms (19.04%), Space: 139.9 MB (21.22%) - LeetHub

+++ 2374-node-with-highest-edge-score/2374-node-with-highest-edge-score.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int edgeScore(vector<int>& edges) {
-        unordered_map<long,long> edgeScores;
+        unordered_map<int,long> edgeScores;
         long maxEdgeScore = 0;
         int ans;
         int n = edges.size();
